Revert transformation to ES modules
- not intended for v5
- move index.ts to index-module.ts
- create new index.ts that provides a CommonJS default export
- clarify that this package type is "commonjs"
- change tests to use module-based index
- add test for CommonJS import
- npm update
