+++ .mocharc.json
@@ -1,5 +1,5 @@
 {
     "extension": ["ts"],
-    "spec": ["test/*.test.ts"],
+    "spec": ["test/*.test.ts", "test/*.test.cjs"],
     "require": "ts-node/register"
 }
