+++ test/mock/service-auth-mock.ts
@@ -17,7 +17,7 @@
  * along with Dicoogle.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-import dicoogleClient from '../../src';
+import dicoogleClient from '../../src/index-module';
 import nock from 'nock';
 
 let nockDone = false;
